Edited Car and Rental entity classes and corresponding repository classes

+++ backend/src/main/java/no/ntnu/project/group4/webapp/repositories/CarRepository.java
@@ -9,5 +9,5 @@ import no.ntnu.project.group4.webapp.models.Car;
  * Repository interface for SQL access to our database table.
  */
 @Repository
-public interface CarRepository extends CrudRepository<Car, Integer> {
+public interface CarRepository extends CrudRepository<Car, Long> {
 }
